Make Parser part of the HasAnalysis
