+++ ouroboros-consensus-cardano/tools/db-analyser/HasAnalysis.hs
@@ -7,6 +7,7 @@ module HasAnalysis (
   ) where
 
 import           Data.Map.Strict (Map)
+import           Options.Applicative
 
 import           Ouroboros.Consensus.Block
 import           Ouroboros.Consensus.Node.ProtocolInfo
@@ -18,6 +19,7 @@ import           Ouroboros.Consensus.Storage.ChainDB.Serialisation (SizeInBytes)
 
 class GetPrevHash blk => HasAnalysis blk where
     data Args blk
+    argsParser      :: proxy blk -> Parser (Args blk)
     mkProtocolInfo  :: Args blk -> IO (ProtocolInfo IO blk)
     countTxOutputs  :: blk -> Int
     blockHeaderSize :: blk -> SizeInBytes
